apps/web/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Django==5.1.4
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-environ==0.12.0
django-js-asset==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kavenegar==1.1.2
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
